+++ TTS/src/core.py
@@ -12,9 +12,9 @@ from huggingface_hub import snapshot_download
 
 from .model.dvae import DVAE
 from .model.audio_generator import AudioGenerator
-from .utils.gpu_utils import select_device
-from .utils.inference_utils import count_invalid_characters, detect_language, apply_character_map, apply_half2full_map, HomophonesReplacer
-from .utils.io_utils import get_latest_modified_file
+from .utils.gpu import select_device
+from .utils.inference import count_invalid_characters, detect_language, apply_character_map, apply_half2full_map, HomophonesReplacer
+from .utils.io import get_latest_modified_file
 from .inference.api import refine_text, inference_code
 from .utils.download import check_all_assets, download_all_assets