+++ TTS/src/model/audio_generator.py
@@ -2,15 +2,16 @@ import os
 os.environ["TOKENIZERS_PARALLELISM"] = "false"
 import logging
 from tqdm import tqdm
-from einops import rearrange
 from transformers.cache_utils import Cache
-
+from transformers import LlamaModel, LlamaConfig
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.nn.utils.parametrize as P
 from torch.nn.utils.parametrizations import weight_norm
-from transformers import LlamaModel, LlamaConfig
+from typing import Union
+
+from ..utils.io import del_all
 
 
 class LlamaMLP(nn.Module):
@@ -29,46 +30,56 @@ class LlamaMLP(nn.Module):
 
 
 class AudioGenerator(nn.Module):
-    def __init__(self, config, num_audio_tokens, num_text_tokens, num_vq=4):
+    def __init__(self, gpt_config, num_audio_tokens, num_text_tokens, num_vq=4, device="cpu"):
         super().__init__()
 
         self.logger = logging.getLogger(__name__)
-        self.model = self.build_model(config)
-        self.model_dim = self.model.config.hidden_size
-
+        self.device = device
+        self.device_gpt = device if "mps" not in str(device) else "cpu"
         self.num_vq = num_vq
-        self.emb_code = nn.ModuleList([nn.Embedding(num_audio_tokens, self.model_dim) for i in range(self.num_vq)])
-        self.emb_text = nn.Embedding(num_text_tokens, self.model_dim)
-        self.head_text = weight_norm(nn.Linear(self.model_dim, num_text_tokens, bias=False), name='weight')
-        self.head_code = nn.ModuleList(
-            [weight_norm(nn.Linear(self.model_dim, num_audio_tokens, bias=False), name='weight') for i in
-             range(self.num_vq)])
 
-    def build_model(self, config):
+        self.gpt = self.build_model(gpt_config, self.device_gpt)
+        self.model_dim = self.gpt.config.hidden_size
+        self.emb_code = nn.ModuleList([nn.Embedding(num_audio_tokens, self.model_dim, device=self.device_gpt)
+                                       for _ in range(num_vq)])
+        self.emb_text = nn.Embedding(num_text_tokens, self.model_dim, device=self.device_gpt)
+
+        self.head_text = weight_norm(nn.Linear(self.model_dim, num_text_tokens, bias=False, device=device), name='weight')
+        self.head_code = nn.ModuleList([weight_norm(nn.Linear(self.model_dim, num_audio_tokens, bias=False, device=device), name='weight')
+                                        for _ in range(self.num_vq)])
+
+    @staticmethod
+    def build_model(config, device):
         configuration = LlamaConfig(**config)
         model = LlamaModel(configuration)
         del model.embed_tokens
 
-        return model
+        return model.to(device)
 
-    def get_emb(self, input_ids, text_mask, **kwargs):
-        emb_text = self.emb_text(input_ids[text_mask][:, 0])
+    def get_emb(self, input_ids, text_mask):
+        emb_text = self.emb_text(input_ids[text_mask][:, 0].to(self.device_gpt))
 
-        emb_code = [self.emb_code[i](input_ids[~text_mask][:, i]) for i in range(self.num_vq)]
+        text_mask_inv = ~text_mask
+        masked_input_ids = input_ids[text_mask_inv].to(self.device_gpt)
+        del text_mask_inv
+
+        emb_code = [self.emb_code[i](masked_input_ids[:, i]) for i in range(self.num_vq)]
         emb_code = torch.stack(emb_code, 2).sum(2)
 
         emb = torch.zeros((input_ids.shape[:-1]) + (emb_text.shape[-1],), device=emb_text.device, dtype=emb_text.dtype)
         emb[text_mask] = emb_text
         emb[~text_mask] = emb_code.to(emb.dtype)
 
+        del emb_text, emb_code
+
         return emb
 
     def prepare_inputs_for_generation(self, input_ids, past_key_values=None, attention_mask=None, inputs_embeds=None, cache_position=None, **kwargs):
         # With static cache, the `past_key_values` is None
-        # TODO: standardize interface for the different Cache classes and remove of this if
+        # TODO joao: standardize interface for the different Cache classes and remove of this if
         has_static_cache = False
         if past_key_values is None:
-            past_key_values = getattr(self.model.layers[0].self_attn, "past_key_value", None)
+            past_key_values = getattr(self.gpt.layers[0].self_attn, "past_key_value", None)
             has_static_cache = past_key_values is not None
 
         past_length = 0
@@ -81,7 +92,7 @@ class AudioGenerator(nn.Module):
                     else None
                 )
                 cache_length = past_length if max_cache_length is None else torch.min(max_cache_length, past_length)
-            # TODO: remove this `else` after `generate` prioritizes `Cache` objects
+            # TODO joao: remove this `else` after `generate` prioritizes `Cache` objects
             else:
                 cache_length = past_length = past_key_values[0][0].shape[2]
                 max_cache_length = None
@@ -128,20 +139,18 @@ class AudioGenerator(nn.Module):
         if has_static_cache:
             past_key_values = None
 
-        model_inputs.update(
-            {
-                "position_ids": position_ids,
-                "cache_position": cache_position,
-                "past_key_values": past_key_values,
-                "use_cache": kwargs.get("use_cache"),
-                "attention_mask": attention_mask,
-            }
-        )
+        model_inputs.update({
+            "position_ids": position_ids,
+            "cache_position": cache_position,
+            "past_key_values": past_key_values,
+            "use_cache": kwargs.get("use_cache"),
+            "attention_mask": attention_mask})
+
         return model_inputs
 
-    def generate(self, emb, inputs_ids, temperature, eos_token, attention_mask=None, max_new_token=2048, min_new_token=0,
-                 LogitsWarpers=[], LogitsProcessors=[], infer_text=False, return_attn=False,
-                 return_hidden=False, stream=False):
+    def generate(self, emb, inputs_ids, temperature, eos_token: Union[int, torch.Tensor], attention_mask=None,
+                 max_new_token=2048, min_new_token=0, LogitsWarpers=[], LogitsProcessors=[],
+                 infer_text=False, return_attn=False, return_hidden=False, stream=False):
         with torch.no_grad():
             attentions = []
             hiddens = []
@@ -150,10 +159,10 @@ class AudioGenerator(nn.Module):
                                                                   dtype=torch.long)
             finish = torch.zeros(inputs_ids.shape[0], device=inputs_ids.device).bool()
 
-            temperature = temperature[None].expand(inputs_ids.shape[0], -1)
-            temperature = rearrange(temperature, "b n -> (b n) 1")
+            temperature = temperature.unsqueeze_(0).expand(inputs_ids.shape[0], -1).contiguous().view(-1, 1)
+            # temperature = rearrange(temperature, "b n -> (b n) 1")
 
-            attention_mask_cache = torch.ones((inputs_ids.shape[0], inputs_ids.shape[1] + max_new_token,),
+            attention_mask_cache = torch.ones((inputs_ids.shape[0], inputs_ids.shape[1] + max_new_token),
                                               dtype=torch.bool, device=inputs_ids.device)
             if attention_mask is not None:
                 attention_mask_cache[:, :attention_mask.shape[1]] = attention_mask
@@ -162,28 +171,24 @@ class AudioGenerator(nn.Module):
                 past_key_values = None
 
                 for i in range(max_new_token):
-                    pbar.update(1)
-                    model_input = self.prepare_inputs_for_generation(
-                        inputs_ids,
-                        past_key_values,
-                        attention_mask_cache[:, :inputs_ids.shape[1]],
-                        use_cache=True,
-                    )
-
+                    model_input = self.prepare_inputs_for_generation(inputs_ids, past_key_values, attention_mask_cache[:, :inputs_ids.shape[1]], use_cache=True)
                     if i == 0:
                         model_input['inputs_embeds'] = emb
                     else:
+                        inputs_ids_emb = model_input['input_ids'].to(self.device_gpt)
                         if infer_text:
-                            model_input['inputs_embeds'] = self.emb_text(model_input['input_ids'][:, :, 0])
+                            model_input['inputs_embeds'] = self.emb_text(inputs_ids_emb[:, :, 0])
                         else:
-                            code_emb = [self.emb_code[i](model_input['input_ids'][:, :, i]) for i in range(self.num_vq)]
+                            code_emb = [self.emb_code[i](inputs_ids_emb[:, :, i]) for i in range(self.num_vq)]
                             model_input['inputs_embeds'] = torch.stack(code_emb, 3).sum(3)
+                        del inputs_ids_emb, model_input['input_ids']
 
-                    model_input['input_ids'] = None
-                    outputs = self.model.forward(**model_input, output_attentions=return_attn)
-                    del model_input
+                    outputs = self.gpt.forward(attention_mask=model_input["attention_mask"].to(self.device_gpt),
+                                               position_ids=model_input["position_ids"].to(self.device_gpt),
+                                               past_key_values=model_input["past_key_values"], inputs_embeds=model_input['inputs_embeds'].to(self.device_gpt), use_cache=model_input['use_cache'], output_attentions=return_attn, cache_position=model_input['cache_position'].to(self.device_gpt),)
+                    del_all(model_input)
                     attentions.append(outputs.attentions)
-                    hidden_states = outputs[0]
+                    hidden_states = outputs[0].to(self.device)  # 🐻
                     past_key_values = outputs.past_key_values
                     del outputs
                     if return_hidden:
@@ -198,8 +203,14 @@ class AudioGenerator(nn.Module):
                     logits = logits[:, -1].float()
 
                     if not infer_text:
-                        logits = rearrange(logits, "b c n -> (b n) c")
-                        logits_token = rearrange(inputs_ids[:, start_idx:], "b c n -> (b n) c")
+                        # logits = rearrange(logits, "b c n -> (b n) c")
+                        logits = logits.permute(0, 2, 1)
+                        logits = logits.reshape(-1, logits.size(2))
+                        # logits_token = rearrange(inputs_ids[:, start_idx:], "b c n -> (b n) c")
+                        inputs_ids_sliced = inputs_ids[:, start_idx:].permute(0, 2, 1)
+                        logits_token = inputs_ids_sliced.reshape(
+                            inputs_ids_sliced.size(0) * inputs_ids_sliced.size(1), -1,
+                        )
                     else:
                         logits_token = inputs_ids[:, start_idx:, 0]
 
@@ -220,10 +231,11 @@ class AudioGenerator(nn.Module):
 
                     del logits
 
-                    idx_next = torch.multinomial(scores, num_samples=1)
+                    idx_next = torch.multinomial(scores, num_samples=1).to(finish.device)
 
                     if not infer_text:
-                        idx_next = rearrange(idx_next, "(b n) 1 -> b n", n=self.num_vq)
+                        # idx_next = rearrange(idx_next, "(b n) 1 -> b n", n=self.num_vq)
+                        idx_next = idx_next.view(-1, self.num_vq)
                         finish_or = (idx_next == eos_token).any(1)
                         finish |= finish_or
                         del finish_or
@@ -252,9 +264,11 @@ class AudioGenerator(nn.Module):
                             'attentions': attentions,
                             'hiddens': y_hiddens,
                         }
+
                     if finish.all():
                         pbar.update(max_new_token - i - 1)
                         break
+                    pbar.update(1)
 
             inputs_ids = [inputs_ids[idx, start_idx: start_idx + i] for idx, i in enumerate(end_idx.int())]
             inputs_ids = [i[:, 0] for i in inputs_ids] if infer_text else inputs_ids